requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
netifaces2==0.0.22
numpy==2.2.3
oauth2client==4.1.3
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyDrive==1.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scapy==2.6.1
six==1.17.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
